+++ ElvUI_EltreumUI/Core/Options/Changelog.lua
@@ -14,35 +14,30 @@ function ElvUI_EltreumUI:Changelog()
 	ElvUI_EltreumUI.Options.args.changelog.args.added = E.Libs.ACH:Group(E:TextGradient("Added", 0.50, 0.70, 1, 0.67, 0.95, 1), nil, 3)
 	ElvUI_EltreumUI.Options.args.changelog.args.added.inline = true
 	ElvUI_EltreumUI.Options.args.changelog.args.added.args.description = E.Libs.ACH:Description([[
-Added an option to disable the mouseover of the Damage Meter
-Added an option to disable the coloring of Rare Nameplates
-Added an option to disable Mirror Timer skins
-Added [name:eltruism:gradientreverse] tag
+Added a new tag [eltruism:status] that combines the dead and dc icons, check note below
+Added an option to disable replacing Damage Meter icons
 ]], 3, "small", nil, nil, nil, nil, "full")
 
 	--updated
 	ElvUI_EltreumUI.Options.args.changelog.args.updated = E.Libs.ACH:Group(E:TextGradient("Updated", 0.50, 0.70, 1, 0.67, 0.95, 1), nil, 3)
 	ElvUI_EltreumUI.Options.args.changelog.args.updated.inline = true
 	ElvUI_EltreumUI.Options.args.changelog.args.updated.args.description = E.Libs.ACH:Description([[
-Updated Damage Meter skin to add Acidweb's fix for the clamping
-Updated some Datatext options that are no longer used
-Updated the nameplate backdrop remover to work again
-Updated several tags to solve issues with secrets
-Updated Retail talent scale (thanks Sash)
-Updated German locale by Dlarge
-Updated Windtools profile
+Updated Actionbar Glows to once again work in TBC and Wrath
+Updated Damage Meter skin to disable embed by default
+Updated Sockets to once again work in Wrath and TBC
+Updated Portrait Fix to add more models
 --]], 5, "small", nil, nil, nil, nil, "full")
 
 	--fixed
 	ElvUI_EltreumUI.Options.args.changelog.args.fixed = E.Libs.ACH:Group(E:TextGradient("Fixed", 0.50, 0.70, 1, 0.67, 0.95, 1), nil, 4)
 	ElvUI_EltreumUI.Options.args.changelog.args.fixed.inline = true
 	ElvUI_EltreumUI.Options.args.changelog.args.fixed.args.description = E.Libs.ACH:Description([[
-Fixed an issue where the Holy Paladin spec icon had a Devourer Demon Hunter overlayed
-Fixed several issues related to PVP instances after the hotfix made them secret
-Fixed an issue where the Leave Vehicle Button no longer had the Texture
-Fixed an issue where some Portrait options were not applying
-Fixed an issue related to using transparent Power
-Fixed an issue with Borders and secrets
-Fixed issues related to chat in Retail
+Fixed an issue with the Quest Hiding not working correctly during Delve and Mini bosses
+Fixed an issue where the Damage Meter icons could error if the Pack was missing
+Fixed an issue where the BetterCooldownManager popup wouldn't be dismissed
+Fixed an issue where all nameplate bars would be colored gradient
+Fixed an issue in TBC where loot quality filter could error
+Fixed issues with the Compatibility Check
+Fixed an issue with the AFK screen
 ]], 7, "small", nil, nil, nil, nil, "full")
 end