+++ Changelog.md
@@ -1,6 +1,21 @@
 ### 🌌 Next Version
 The future is uncertain
 ---
+### 5.0.3
+Added a new tag [eltruism:status] that combines the dead and dc icons, check note below  
+Added an option to disable replacing Damage Meter icons  
+Updated Actionbar Glows to once again work in TBC and Wrath  
+Updated Damage Meter skin to disable embed by default  
+Updated Sockets to once again work in Wrath and TBC  
+Updated Portrait Fix to add more models  
+Fixed an issue with the Quest Hiding not working correctly during Delve and Mini bosses  
+Fixed an issue where the Damage Meter icons could error if the Pack was missing  
+Fixed an issue where the BetterCooldownManager popup wouldn't be dismissed  
+Fixed an issue where all nameplate bars would be colored gradient  
+Fixed an issue in TBC where loot quality filter could error  
+Fixed issues with the Compatibility Check  
+Fixed an issue with the AFK screen  
+---
 ### 5.0.2
 Added an option to disable the mouseover of the Damage Meter  
 Added an option to disable the coloring of Rare Nameplates